+++ src/main/java/com/chilicode/JFroniusTBot/model/Inverter.java
@@ -10,57 +10,57 @@ public class Inverter {
 
     @SerializedName("DT")
     @Expose
-    private Integer dt;
+    private Float dt;
     @SerializedName("E_Day")
     @Expose
-    private Integer eDay;
+    private Float eDay;
     @SerializedName("E_Total")
     @Expose
-    private Integer eTotal;
+    private Float eTotal;
     @SerializedName("E_Year")
     @Expose
-    private Integer eYear;
+    private Float eYear;
     @SerializedName("P")
     @Expose
-    private Integer p;
+    private Float p;
 
-    public Integer getDt() {
+    public Float getDt() {
         return dt;
     }
 
-    public void setDt(Integer dt) {
+    public void setDt(Float dt) {
         this.dt = dt;
     }
 
-    public Integer getEDay() {
+    public Float getEDay() {
         return eDay;
     }
 
-    public void setEDay(Integer eDay) {
+    public void setEDay(Float eDay) {
         this.eDay = eDay;
     }
 
-    public Integer getETotal() {
+    public Float getETotal() {
         return eTotal;
     }
 
-    public void setETotal(Integer eTotal) {
+    public void setETotal(Float eTotal) {
         this.eTotal = eTotal;
     }
 
-    public Integer getEYear() {
+    public Float getEYear() {
         return eYear;
     }
 
-    public void setEYear(Integer eYear) {
+    public void setEYear(Float eYear) {
         this.eYear = eYear;
     }
 
-    public Integer getP() {
+    public Float getP() {
         return p;
     }
 
-    public void setP(Integer p) {
+    public void setP(Float p) {
         this.p = p;
     }
